+++ package.json
@@ -108,6 +108,7 @@
     "jest-config": "28.1.1",
     "jest-junit": "13.2.0",
     "jest-mock-random": "1.1.1",
+    "jest-serializer-html": "^7.1.0",
     "js-sha256": "0.9.0",
     "koa": "2.13.4",
     "koa-router": "10.1.1",
