+++ tsconfig.test.json
@@ -20,5 +20,6 @@
     { "path": "./packages/apollo-server-lambda/src/__tests__/" },
     { "path": "./packages/apollo-server-micro/src/__tests__/" },
     { "path": "./packages/apollo-server-plugin-response-cache/src/__tests__/" },
+    { "path": "./packages/apollo-server-core/src/plugin/landingPage/__tests__/" },
   ]
 }
